feat: chore: addTags and setting folder renamed
