+++ routes/mainRoutes.js
@@ -10,7 +10,6 @@ import PermissionStack from './permissionRoutes';
 import { SettingScreen, NewUpdateScreen } from '../screens';
 // import AlbumList from '../screens/pages/library/AlbumList';
 import GalleryStack from './galleryRoutes';
-import AddTags from '../screens/pages/AddTags';
 
 const Stack = createStackNavigator();
