+++ screens/addTag/addTagComponents/LitterBottomSearch.js
