+++ screens/addTag/addTagComponents/index.js
@@ -1,0 +1,6 @@
+export { default as LitterCategories } from './LitterCategories';
+export { default as LitterImage } from './LitterImage';
+export { default as LitterPickerWheels } from './LitterPickerWheels';
+export { default as LitterTags } from './LitterTags';
+export { default as LitterBottomSearch } from './LitterBottomSearch';
+export { default as TagsActionButton } from './TagsActionButton';
