+++ screens/index.js
@@ -2,7 +2,7 @@ export { default as UserStatsScreen } from './userStats/UserStatsScreen';
 export { default as GlobalDataScreen } from './globalData/GlobalDataScreen';
 export { default as TeamScreen } from './team/TeamScreen';
 export { default as RankingScreen } from './ranking/RankingScreen';
-export { default as SettingScreen } from './SettingsScreen';
+export { default as SettingScreen } from './setting/SettingsScreen';
 export { default as HomeScreen } from './home/HomeScreen';
 export { default as CameraScreen } from './camera/CameraScreen';
 export { default as GalleryScreen } from './gallery/GalleryScreen';
