+++ screens/setting/SettingsScreen.js
@@ -15,14 +15,14 @@ import {
 } from 'react-native';
 import { getTranslation, TransText } from 'react-native-translation';
 import Icon from 'react-native-vector-icons/Ionicons';
-import * as actions from '../actions';
+import * as actions from '../../actions';
 import { connect } from 'react-redux';
-import { Body, SubTitle, Title, Header, Colors, Caption } from './components';
+import { Body, SubTitle, Title, Header, Colors, Caption } from '../components';
 
 const SCREEN_WIDTH = Dimensions.get('window').width;
 const SCREEN_HEIGHT = Dimensions.get('window').height;
 
-import SettingsComponent from './pages/SettingsComponent';
+import SettingsComponent from './settingComponents/SettingsComponent';
 
 class SettingsScreen extends Component {
     constructor(props) {
